+++ src/utils.rs
@@ -11,50 +11,13 @@
 use std::cmp::Ordering;
 
 use syntax::ast::{self, Visibility, Attribute, MetaItem, MetaItemKind};
-use syntax::codemap::{CodeMap, Span, BytePos};
+use syntax::codemap::BytePos;
 use syntax::abi;
 
 use Indent;
-use comment::FindUncommented;
 use rewrite::{Rewrite, RewriteContext};
 
 use SKIP_ANNOTATION;
-
-pub trait CodeMapSpanUtils {
-    fn span_after(&self, original: Span, needle: &str) -> BytePos;
-    fn span_after_last(&self, original: Span, needle: &str) -> BytePos;
-    fn span_before(&self, original: Span, needle: &str) -> BytePos;
-}
-
-impl CodeMapSpanUtils for CodeMap {
-    #[inline]
-    fn span_after(&self, original: Span, needle: &str) -> BytePos {
-        let snippet = self.span_to_snippet(original).unwrap();
-        let offset = snippet.find_uncommented(needle).unwrap() + needle.len();
-
-        original.lo + BytePos(offset as u32)
-    }
-
-    #[inline]
-    fn span_after_last(&self, original: Span, needle: &str) -> BytePos {
-        let snippet = self.span_to_snippet(original).unwrap();
-        let mut offset = 0;
-
-        while let Some(additional_offset) = snippet[offset..].find_uncommented(needle) {
-            offset += additional_offset + needle.len();
-        }
-
-        original.lo + BytePos(offset as u32)
-    }
-
-    #[inline]
-    fn span_before(&self, original: Span, needle: &str) -> BytePos {
-        let snippet = self.span_to_snippet(original).unwrap();
-        let offset = snippet.find_uncommented(needle).unwrap();
-
-        original.lo + BytePos(offset as u32)
-    }
-}
 
 // Computes the length of a string's last line, minus offset.
 #[inline]
